add crashes support to puppet app
