+++ Assets/Puppet/PuppetCrashes.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 6a322f155379d304eb7a972afb711bc9
+timeCreated: 1534262520
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
