+++ Assets/Puppet/PuppetOther.cs
@@ -1,8 +1,7 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 //
 // Licensed under the MIT license.
 
-using Microsoft.AppCenter.Unity.Crashes;
 using Microsoft.AppCenter.Unity.Distribute;
 using Microsoft.AppCenter.Unity.Push;
 using System.Collections;
@@ -13,7 +12,6 @@ public class PuppetOther : MonoBehaviour
 {
     public Toggle DistributeEnabled;
     public Toggle PushEnabled;
-    public Toggle CrashesEnabled;
 
     void OnEnable()
     {
@@ -30,10 +28,6 @@ public class PuppetOther : MonoBehaviour
         {
             return true;
         };
-        Crashes.IsEnabledAsync().ContinueWith(task =>
-        {
-            CrashesEnabled.isOn = task.Result;
-        });
     }
 
     public void SetPushEnabled(bool enabled)
@@ -60,18 +54,5 @@ public class PuppetOther : MonoBehaviour
         var isEnabled = Distribute.IsEnabledAsync();
         yield return isEnabled;
         DistributeEnabled.isOn = isEnabled.Result;
-    }
-
-    public void SetCrashesEnabled(bool enabled)
-    {
-        StartCoroutine(SetCrashesEnabledCoroutine(enabled));
-    }
-
-    private IEnumerator SetCrashesEnabledCoroutine(bool enabled)
-    {
-        yield return Crashes.SetEnabledAsync(enabled);
-        var isEnabled = Crashes.IsEnabledAsync();
-        yield return isEnabled;
-        CrashesEnabled.isOn = isEnabled.Result;
     }
 }
